Dtos y Resource

+++ s4_dispositivos-back/src/main/java/co/edu/uniandes/csw/dispositivos/entities/CalificacionEntity.java
@@ -47,7 +47,7 @@ public class CalificacionEntity extends BaseEntity implements Serializable {
     /**
      * @return the calificacionNumerica
      */
-    public long getCalificacionNumerica() {
+    public int getCalificacionNumerica() {
         return calificacionNumerica;
     }
